+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added `blueprint test --ui` command
 
+### Changed
+
+- Changed `@ton/sandbox` peer dependency to `>=0.40.0`
+
 ## [0.41.0] - 2025-09-23
 
 ### Added